fix: address app icon badge review feedback
