+++ Sources/cmuxApp.swift
@@ -4372,12 +4372,15 @@ enum AppIconLaunchState {
 enum AppIconSettings {
     static let modeKey = "appIconMode"
     static let defaultMode: AppIconMode = .automatic
-    private static let dockTileIconDidChangeNotification = Notification.Name("com.cmuxterm.appIconDidChange")
+    fileprivate static let dockTileIconDidChangeNotification = Notification.Name("com.cmuxterm.appIconDidChange")
+    @MainActor
     private static var liveEnvironmentProvider: () -> Environment = { .live() }
+    @MainActor
     private static var runtimeBaseIcon: NSImage?
+    @MainActor
     private static var runtimeBadgeLabel: String?
 
-    private static func isRunningUnderXCTest(_ env: [String: String] = ProcessInfo.processInfo.environment) -> Bool {
+    fileprivate static func isRunningUnderXCTest(_ env: [String: String] = ProcessInfo.processInfo.environment) -> Bool {
         if env["XCTestConfigurationFilePath"] != nil { return true }
         if env["XCTestBundlePath"] != nil { return true }
         if env["XCTestSessionIdentifier"] != nil { return true }
@@ -4391,10 +4394,10 @@ enum AppIconSettings {
     struct Environment {
         let isApplicationFinishedLaunching: () -> Bool
         let imageForMode: (AppIconMode) -> NSImage?
-        let setApplicationIconImage: (NSImage) -> Void
-        let startAppearanceObservation: () -> Void
-        let stopAppearanceObservation: () -> Void
-        let notifyDockTilePlugin: () -> Void
+        let setApplicationIconImage: @MainActor (NSImage) -> Void
+        let startAppearanceObservation: @MainActor () -> Void
+        let stopAppearanceObservation: @MainActor () -> Void
+        let notifyDockTilePlugin: @MainActor () -> Void
 
         static func live() -> Self {
             Self(
@@ -4435,6 +4438,7 @@ enum AppIconSettings {
         return mode
     }
 
+    @MainActor
     static func applyIcon(_ mode: AppIconMode, environment: Environment? = nil) {
         let environment = environment ?? liveEnvironmentProvider()
         // Tahoe can crash or wedge when app icon work runs during App.init(),
@@ -4445,6 +4449,7 @@ enum AppIconSettings {
         switch mode {
         case .automatic:
             environment.startAppearanceObservation()
+            environment.notifyDockTilePlugin()
         case .light:
             environment.stopAppearanceObservation()
             guard let icon = environment.imageForMode(.light) else { return }
@@ -4454,10 +4459,9 @@ enum AppIconSettings {
             guard let icon = environment.imageForMode(.dark) else { return }
             setRuntimeBaseIcon(icon, environment: environment)
         }
-
-        environment.notifyDockTilePlugin()
     }
 
+    @MainActor
     static func updateRuntimeBadgeLabel(_ label: String?, environment: Environment? = nil) {
         runtimeBadgeLabel = AppIconBadgeRenderer.normalizedBadgeLabel(label)
         let environment = environment ?? liveEnvironmentProvider()
@@ -4468,125 +4472,29 @@ enum AppIconSettings {
         environment.notifyDockTilePlugin()
     }
 
+    @MainActor
     static func setRuntimeBaseIcon(_ icon: NSImage, environment: Environment? = nil) {
         runtimeBaseIcon = icon
         let environment = environment ?? liveEnvironmentProvider()
         environment.setApplicationIconImage(runtimeIcon(for: icon))
+        environment.notifyDockTilePlugin()
     }
 
+    @MainActor
     private static func runtimeIcon(for baseIcon: NSImage) -> NSImage {
-        let badgeLabel = runtimeBadgeLabel ?? AppIconBadgeRenderer.normalizedBadgeLabel(NSApp?.dockTile.badgeLabel)
-        return AppIconBadgeRenderer.image(baseIcon: baseIcon, badgeLabel: badgeLabel)
+        AppIconBadgeRenderer.image(baseIcon: baseIcon, badgeLabel: runtimeBadgeLabel)
     }
 
+    @MainActor
     static func setLiveEnvironmentProviderForTesting(_ provider: @escaping () -> Environment) {
         liveEnvironmentProvider = provider
     }
 
+    @MainActor
     static func resetLiveEnvironmentProviderForTesting() {
         liveEnvironmentProvider = { .live() }
-    }
-}
-
-enum AppIconBadgeRenderer {
-    static func normalizedBadgeLabel(_ rawLabel: String?) -> String? {
-        guard let label = rawLabel?.trimmingCharacters(in: .whitespacesAndNewlines), !label.isEmpty else {
-            return nil
-        }
-        return label
-    }
-
-    static func image(baseIcon: NSImage, badgeLabel rawBadgeLabel: String?) -> NSImage {
-        guard let badgeLabel = normalizedBadgeLabel(rawBadgeLabel) else {
-            return baseIcon
-        }
-
-        let size = normalizedIconSize(baseIcon.size)
-        let result = NSImage(size: size)
-        result.isTemplate = false
-        result.lockFocus()
-        defer { result.unlockFocus() }
-
-        baseIcon.draw(
-            in: NSRect(origin: .zero, size: size),
-            from: NSRect(origin: .zero, size: baseIcon.size),
-            operation: .sourceOver,
-            fraction: 1
-        )
-        drawBadge(badgeLabel, in: NSRect(origin: .zero, size: size))
-        return result
-    }
-
-    private static func normalizedIconSize(_ size: NSSize) -> NSSize {
-        guard size.width > 0, size.height > 0 else {
-            return NSSize(width: 128, height: 128)
-        }
-        return size
-    }
-
-    private static func drawBadge(_ label: String, in bounds: NSRect) {
-        guard bounds.width > 0, bounds.height > 0 else { return }
-
-        NSGraphicsContext.current?.shouldAntialias = true
-
-        let iconEdge = min(bounds.width, bounds.height)
-        let badgeHeight = max(18, iconEdge * 0.34)
-        let horizontalPadding = badgeHeight * 0.28
-        let maxBadgeWidth = bounds.width * 0.86
-        let font = badgeFont(
-            fitting: label,
-            badgeHeight: badgeHeight,
-            horizontalPadding: horizontalPadding,
-            maxBadgeWidth: maxBadgeWidth
-        )
-        let attributes: [NSAttributedString.Key: Any] = [
-            .font: font,
-            .foregroundColor: NSColor.white,
-        ]
-        let textSize = label.size(withAttributes: attributes)
-        let badgeWidth = min(maxBadgeWidth, max(badgeHeight, ceil(textSize.width + horizontalPadding * 2)))
-        let badgeRect = NSRect(
-            x: bounds.maxX - badgeWidth - iconEdge * 0.02,
-            y: bounds.maxY - badgeHeight - iconEdge * 0.02,
-            width: badgeWidth,
-            height: badgeHeight
-        )
-
-        NSColor.white.setFill()
-        NSBezierPath(
-            roundedRect: badgeRect.insetBy(dx: -badgeHeight * 0.07, dy: -badgeHeight * 0.07),
-            xRadius: badgeHeight * 0.58,
-            yRadius: badgeHeight * 0.58
-        ).fill()
-        NSColor(calibratedRed: 1.0, green: 0.12, blue: 0.16, alpha: 1.0).setFill()
-        NSBezierPath(roundedRect: badgeRect, xRadius: badgeHeight / 2, yRadius: badgeHeight / 2).fill()
-
-        let textRect = NSRect(
-            x: badgeRect.midX - textSize.width / 2,
-            y: badgeRect.midY - textSize.height / 2,
-            width: textSize.width,
-            height: textSize.height
-        )
-        label.draw(in: textRect, withAttributes: attributes)
-    }
-
-    private static func badgeFont(
-        fitting label: String,
-        badgeHeight: CGFloat,
-        horizontalPadding: CGFloat,
-        maxBadgeWidth: CGFloat
-    ) -> NSFont {
-        let baseSize = max(11, badgeHeight * 0.56)
-        let baseFont = NSFont.systemFont(ofSize: baseSize, weight: .bold)
-        let textWidth = label.size(withAttributes: [.font: baseFont]).width
-        let availableWidth = max(1, maxBadgeWidth - horizontalPadding * 2)
-        guard textWidth > availableWidth else {
-            return baseFont
-        }
-        return NSFont.systemFont(
-            ofSize: max(8, floor(baseSize * availableWidth / max(textWidth, 1))),
-            weight: .bold
-        )
+        runtimeBaseIcon = nil
+        runtimeBadgeLabel = nil
     }
 }
 
@@ -4596,6 +4504,7 @@ protocol AppIconAppearanceObservation: AnyObject {
 
 extension NSKeyValueObservation: AppIconAppearanceObservation {}
 
+@MainActor
 final class AppIconAppearanceObserver: NSObject {
     struct Environment {
         let isApplicationFinishedLaunching: () -> Bool
@@ -4605,6 +4514,7 @@ final class AppIconAppearanceObserver: NSObject {
         let currentAppearanceIsDark: () -> Bool?
         let imageForName: (String) -> NSImage?
         let setApplicationIconImage: (NSImage) -> Void
+        let notifyDockTilePlugin: () -> Void
 
         static func live() -> Self {
             Self(
@@ -4640,6 +4550,15 @@ final class AppIconAppearanceObserver: NSObject {
                 },
                 setApplicationIconImage: { icon in
                     NSApplication.shared.applicationIconImage = icon
+                },
+                notifyDockTilePlugin: {
+                    guard !AppIconSettings.isRunningUnderXCTest() else { return }
+                    DistributedNotificationCenter.default().postNotificationName(
+                        AppIconSettings.dockTileIconDidChangeNotification,
+                        object: nil,
+                        userInfo: nil,
+                        deliverImmediately: true
+                    )
                 }
             )
         }
@@ -4707,7 +4626,7 @@ final class AppIconAppearanceObserver: NSObject {
             setApplicationIconImage: environment.setApplicationIconImage,
             startAppearanceObservation: {},
             stopAppearanceObservation: {},
-            notifyDockTilePlugin: {}
+            notifyDockTilePlugin: environment.notifyDockTilePlugin
         ))
         lastAppliedImageName = imageName
     }
