+++ cmuxTests/KeyboardShortcutSettingsFileStoreStartupTests.swift
@@ -7,6 +7,7 @@ import AppKit
 @testable import cmux
 #endif
 
+@MainActor
 final class KeyboardShortcutSettingsFileStoreStartupTests: XCTestCase {
     private var originalSettingsFileStore: KeyboardShortcutSettingsFileStore!
     private let settingsFileBackupsDefaultsKey = "cmux.settingsFile.backups.v1"
